fix: harden Lotus workflow command discovery
